% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{combTMB-methods}
\alias{combTMB-methods}
\alias{print.combTMB}
\alias{fitted.combTMB}
\alias{family.combTMB}
\alias{nobs.combTMB}
\alias{df.residual.combTMB}
\alias{logLik.combTMB}
\alias{vcov.combTMB}
\alias{print.vcov.combTMB}
\alias{model.matrix.combTMB}
\alias{fixef.combTMB}
\alias{coefDisp}
\title{Method for 'combTMB' objects}
\usage{
\method{print}{combTMB}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{fitted}{combTMB}(object, ...)

\method{family}{combTMB}(object, ...)

\method{nobs}{combTMB}(object, ...)

\method{df.residual}{combTMB}(object, ...)

\method{logLik}{combTMB}(object, ...)

\method{vcov}{combTMB}(object, ...)

\method{print}{vcov.combTMB}(x, ...)

\method{model.matrix}{combTMB}(object, ...)

\method{fixef}{combTMB}(object, ...)

coefDisp(object, ...)
}
\arguments{
\item{x}{an object of class \code{combTMB}}

\item{digits}{minimal number of \emph{significant} digits, see
\code{\link[base]{print.default}}}

\item{\dots}{currently not used}

\item{object}{an object of class \code{combTMB}}
}
\value{
.
}
\description{
Method for 'combTMB' objects
}
\details{
The print method for \code{fixef.combTMB} object \emph{only displays non-trivial components}.
}
