% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{combTMB-package}
\alias{combTMB-package}
\alias{_PACKAGE}
\title{combTMB: Combined Models using Template Model Builder}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements an extended random effects approach to model repeated, overdispersed binary and count data (called combined models). Furthermore, fit linear and generalized linear mixed. The adjustment of the models is done by maximum likelihood with 'TMB' (Template Model Builder). A common application is for superdispersed and correlated longitudinal data. See Molenberghs, G., Verbeke, G. and Demetrio, C.G. (2007) \doi{10.1007/s10985-007-9064-y} and Molenberghs et al. (2012)\doi{10.1016/j.jmva.2012.05.005}.
}
\author{
\strong{Maintainer}: Deoclecio Jardim Amorim \email{deocleciojardimamorim@gmail.com} (\href{https://orcid.org/0000-0002-2844-3239}{ORCID})

Authors:
\itemize{
  \item Afranio Marcio Correa Vieira (\href{https://orcid.org/0000-0003-1187-6377}{ORCID})
  \item Clarice Garcia Borges Demetrio (\href{https://orcid.org/0000-0002-3609-178X}{ORCID})
}

}
\keyword{internal}
