% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combTMBcontrol.R
\name{combTMBcontrol}
\alias{combTMBcontrol}
\title{Optimization control options}
\usage{
combTMBcontrol(
  optimizer = c("nlminb", "optim"),
  optim.method = "BFGS",
  eval.max = 2000L,
  iter.max = 2000L,
  maxit = 500L,
  nlminbLoops = 1L,
  newtonLoops = 0L,
  zerodisp_val = log(sqrt(.Machine$double.eps)),
  start_params = NULL,
  lower = NULL,
  upper = NULL,
  gjp = TRUE,
  ...
)
}
\arguments{
\item{optimizer}{optimization with "\code{\link[stats]{nlminb}}" or
"\code{\link[stats]{optim}}".}

\item{optim.method}{by default, the optimizer performs minimization with
the `BFGS` method. If the lower and upper limits are declared the indicated
 method is the `L-BFGS-B`.}

\item{eval.max}{maximum number of evaluations of the objective function
allowed.}

\item{iter.max}{maximum number of iterations allowed.}

\item{maxit}{the maximum number of iterations for the "\code{\link[stats]{optim}}" method.}

\item{nlminbLoops}{how many times to run "\code{\link[stats]{nlminb}}" optimization.
Sometimes restarting the optimizer at the previous best values aids
convergence. If the maximum gradient is still too large,
try increasing this to `2`.}

\item{newtonLoops}{how many Newton optimization steps to try with
"\code{\link[stats]{optimHess}}" after running "\code{\link[stats]{nlminb}}".
 Sometimes aids convergence.}

\item{zerodisp_val}{value of the dispersion parameter when \code{dformula=~0} is specified.}

\item{start_params}{the user can specify initial values for the parameters.}

\item{lower}{an optional named list of lower bounds within the optimization.}

\item{upper}{an optional named list of upper bounds within the optimization.}

\item{gjp}{logical. Passed to `getJointPrecision` in "\code{\link[TMB]{sdreport}}".
Must be `TRUE` to use simulation-based methods. If not needed, setting this
 `FALSE` will reduce object size.}

\item{\dots}{currently not used.}
}
\description{
Optimization control options
}
\author{
Deoclecio Jardim Amorim <deocleciojardimamorim@gmail.com>.
}
