% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dharma_residual.R
\name{dharma_residuals}
\alias{dharma_residuals}
\title{DHARMa residuals}
\usage{
dharma_residuals(object, nsim = 250, seed = 123, plot = TRUE, ...)
}
\arguments{
\item{object}{an \dQuote{combTMB} model}

\item{nsim}{number of simulations to be executed (n > 1). The default is 250,
however it is recommended to run through 1000 simulations}

\item{seed}{random number seed. Guarantees the reproducibility of results}

\item{plot}{logical. Default is `TRUE`, if `FALSE` the graph is not plotted}

\item{...}{Other arguments to pass to [DHARMa::createDHARMa()]}
}
\value{
A list with two objects out_1 and out_2 is invisibly returned. In out_1
a data frame with expected and observed values is returned, if `plot = FALSE`,
you can assign an output to the object and plot the residuals,
for example, with the \pkg{ggplo2} package. In out_2 is the complete output
of the function [DHARMa::createDHARMa()] is returned, which lets us use other
\pkg{DHARMa} tools.
}
\description{
DHARMa residuals
}
\examples{
\donttest{
library(combTMB)
#Poisson-Normal model
m1 <- combTMB(OT ~ Period+(1|Donor), family=poisson(), data=embryos)

#Dharma residuals
if (require("DHARMa", quietly = TRUE)) {
r <- dharma_residuals(m1, nsim=250)
head(r$out_1)
}
}
}
