% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\alias{dispersion.combTMB}
\title{Extract residual standard deviation or dispersion parameter}
\usage{
dispersion(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{combTMB}}

\item{\dots}{currently not used.}
}
\value{
.
}
\description{
Extract residual standard deviation or dispersion parameter
}
