% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{emmeans.combTMB}
\alias{emmeans.combTMB}
\title{Estimated marginal means with the \pkg{emmeans} package with \pkg{comTMB}}
\description{
Methods for using the \pkg{emmeans} package with \pkg{combTMB}. The
\pkg{emmeans} package computes estimated marginal means for the fixed
effects.
}
\references{
\url{https://aosmith.rbind.io/2019/03/25/getting-started-with-emmeans/}
}
