% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mcmc.R
\name{extract_mcmc}
\alias{extract_mcmc}
\title{Extract MCMC samples from an combTMB model fit with tmbstan.}
\usage{
extract_mcmc(object)
}
\arguments{
\item{object}{Output from [tmbstan::tmbstan()] run on the `tmb_obj`
element of an [combTMB::combTMB()] model. E.g.,
`tmbstan::tmbstan(your_model$tmb_obj)`.}
}
\value{
Returns a matrix of parameter samples. Rows correspond to the order
of `your_model$tmb_obj$env$last.par.best`. Columns correspond to
posterior samples.
}
\description{
Extract MCMC samples from an combTMB model fit with tmbstan.
}
