% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extractors}
\alias{extractors}
\alias{getME.combTMB}
\alias{getME}
\title{Extract or obtain generalized components from a glm or glmm model or combined model}
\usage{
\method{getME}{combTMB}(object, name = c("X", "Xd", "Z", "beta", "betad", "b"), ...)
}
\arguments{
\item{object}{an object of class \dQuote{combTMB}}

\item{name}{of the component to be retrieved}

\item{\dots}{currently not used.}
}
\value{
.
}
\description{
Extract or obtain generalized components from a glm or glmm model or combined model
}
\seealso{
\code{\link[lme4]{getME}}
}
