% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family_combTMB}
\alias{family_combTMB}
\alias{cmp}
\alias{gpoisson}
\alias{geometric}
\alias{bbn}
\alias{beta_fam}
\alias{betabinomial}
\alias{poigamma}
\alias{gnb}
\alias{betabernoulli}
\alias{betabernoulli_al}
\title{Family functions for combTMB}
\usage{
cmp(link = "log")

gpoisson(link = "log")

geometric(link = "log")

bbn(link = "log")

beta_fam(link = "logit")

betabinomial(link = "logit")

poigamma(link = "log")

gnb(link = "log")

betabernoulli(link = "logit")

betabernoulli_al(link = "probit")
}
\arguments{
\item{link}{An character link function for the mean
("log", "logit", "probit", "cloglog","inverse" ,"identity", or "sqrt").}
}
\value{
A list with elements `family`, `link`, `linkfun`, `linkinv`, and `mu.eta`.
}
\description{
Family functions for combTMB
}
\details{
For the dispersion model, the log link was used. Denoting the variance as
\eqn{V}, the dispersion parameter as \eqn{\phi=\exp(\eta)}{phi=exp(eta)}
(where \eqn{\eta}{eta} is the linear predictor from the dispersion model),
and the predicted mean as \eqn{\mu}{mu}:
\describe{
\item{gaussian}{(from base R): constant \eqn{V=\phi}{V=phi}.}

\item{Beta}{Beta distribution: parameterization of Cribari-Neto and Zeileis
(2010) (\pkg{betareg} package); \eqn{V=\mu(1-\mu)/(\phi+1)}{V=mu*(1-mu)/(phi+1)}.}

\item{betabinomial}{Beta-binomial distribution: parameterized according
to Morris (1997). \eqn{V=\mu(1-\mu)(n(\phi+n)/(\phi+1))}{V=mu*(1-mu)*(n*(phi+n)/(phi+1))}.}

\item{poigamma}{Poisson-Gamma: a combined model  for count data
that accommodates overdispersion and clustering into two separate random
effects sets, gamma and normal, respectively (Molenberghs et al., 2007)).
Negative binomial distribution type II: quadratic parameterization
 (Hardin & Hilbe 2007).
 \eqn{V=\mu(1+\mu/\phi) = \mu+\mu^2/\phi}{V=mu*(1+mu/phi) = mu+mu^2/phi}.}

\item{betabernoulli}{Beta-Bernoulli: a combined model model for binary data
that accommodates overdispersion and clustering into two separate random
effects sets, Beta and normal, respectively
(Molenberghs et al., 2010 and Molenberghs et al., 2012)).}

\item{betabernoulli_al}{Beta-Bernoulli: a combined model for binary data
that accommodates overdispersion and clustering into two separate random
effects sets, Beta and normal, respectively
(Molenberghs et al., 2010 and Molenberghs et al., 2012)).}

}
}
\examples{
library(combTMB)
cmp(link = "log")
gpoisson(link = "log")
geometric(link = "log")
bbn(link = "log")
beta_fam(link = "logit")
betabinomial(link = "logit")
poigamma(link = "log")
gnb(link = "log")
betabernoulli(link = "logit")
betabernoulli_al(link = "probit")
}
\references{
\itemize{
\item Hardin JW & Hilbe JM (2007). "Generalized linear models and extensions."
 Stata Press.

\item Ferrari SLP, Cribari-Neto F (2004). "Beta Regression for Modelling
Rates and Proportions." \emph{J. Appl. Stat.}  31(7), 799-815.

\item Molenberghs, G., Verbeke, G., and Demetrio, C. G. (2007). An extended
random-effects approach to modeling repeated, overdispersed count data.
\emph{Lifetime data analysis}, 13(4), 513-531.

\item Molenberghs, G., Verbeke, G., Demetrio, C. G., and Vieira, A. M. (2010).
A family of generalized linear models for repeated measures with normal
and conjugate random effects. \emph{Statistical science}, 25(3), 325-347.

\item Molenberghs, G., Verbeke, G., Iddi, S., and Demetrio, C. G. (2012).
A combined beta and normal random-effects model for repeated,
overdispersed binary and binomial data.
\emph{Journal of Multivariate Analysis}, 111, 94-109.

\item Morris  W (1997). "Disentangling Effects of Induced Plant Defenses and
 Food Quantity on Herbivores by Fitting Nonlinear Models."
 \emph{American Naturalist} 150:299-327.
}
}
