% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{lrtests-combTMB}
\alias{lrtests-combTMB}
\alias{anova.combTMB}
\title{Likelihood ratio tests for nested combTMB models}
\usage{
\method{anova}{combTMB}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{combTMB}, a result of call
\code{\link{combTMB}(...)}}

\item{\dots}{currently not used}
}
\value{
an object of class \dQuote{combTMB}, a table with compenents
  for Likelihood ratio (LR) test
}
\description{
Likelihood ratio tests for nested combTMB models
}
\details{
it is the user's responsibility to identify whether the models are nested!
}
\examples{
\donttest{
library(combTMB)
#Poisson-Normal model
m0 <- combTMB(OT ~ 1+(1|Donor), family=poisson(), data=embryos)

#Poisson-Normal model
m1 <- combTMB(OT ~ Period+(1|Donor), family=poisson(), data=embryos)

#LR test for Period
anova(m0, m1)

}

}
