% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partvar_bbn.R
\name{partvar}
\alias{partvar}
\title{Partition of the variance in a bbn family in combTMB}
\usage{
partvar(object, k, rho, ...)
}
\arguments{
\item{object}{an object class \code{combTMB} for which the partition is desired.}

\item{k}{value obtained parameter obtained after adjustment}

\item{rho}{value obtained parameter obtained after adjustment}

\item{\dots}{further arguments passed to or from other methods}
}
\value{
Two data frames in tibble format, with ratio of sources of variation
and sources of variation in which variance is splitted.
}
\description{
In a bbn model, the variance may be split into three terms.
The first component of this decomposition represents the
variability due to randomness and it comes from the underlying Poisson model.
The other two components refer to the variability that is not due
to randomness but is explained by the presence of liability and proneness, respectively.
This function is similar to the one presented in the \pkg{GWRM} package [GWRM::partvar],
but the user needs to inform the values of k and rho.
}
