% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\docType{methods}
\name{ranef.combTMB}
\alias{ranef.combTMB}
\alias{ranef}
\title{Extract Random Effects}
\usage{
\method{ranef}{combTMB}(object, ...)
}
\arguments{
\item{object}{a \code{combTMB} model}

\item{\dots}{Currently not used}
}
\value{
For \code{ranef}, an object of class \code{ranef.combTMB} with two components:
\describe{
  \item{cond}{a list of data frames, containing random effects}
    \item{fullranef}{a data frame with:}
   \describe{
      \item{group}{level of the grouping variable}
      \item{estimate}{value of the random-effects}
      \item{std.error}{standard deviation}
       }
}
}
\description{
Extract random effects from a fitted \code{combTMB} model
}
\note{
To show the complete structure use:
 \code{print(ranef(model),simplify=FALSE)}
}
\examples{
\donttest{
library(combTMB)
#Poisson-Normal model
m1 <- combTMB(OT ~ Period+(1|Donor), family=poisson(), data=embryos)
summary(m1)
r <- ranef(m1)
print(r, simplify=FALSE)
#Extract Donor
r$cond$Donor
}
}
