% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_combTMB}
\alias{simulate_combTMB}
\alias{simulate.combTMB}
\title{Simulate from a combTMB fitted model}
\usage{
\method{simulate}{combTMB}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{combTMB fitted model}

\item{nsim}{number of response lists to simulate. Defaults to 1.}

\item{seed}{random number seed. Guarantees the reproducibility of results}

\item{\dots}{currently not used}
}
\value{
A tibble. The list has length \code{nsim}.
Each simulated vector of observations is the same size as the vector of response
variables in the original data set.In the binomial family case each simulation
is a two-column matrix with success/failure.
}
\description{
Simulate from a combTMB fitted model
}
\details{
In this function, random effects are simulated from their estimated
distribution. In the current version, it is not possible to condition the
estimated random effects.
}
