% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{summary-combTMB}
\alias{summary-combTMB}
\alias{summary.combTMB}
\alias{print.summary.combTMB}
\title{Summary of the combTMB models (individual t-tests)}
\usage{
\method{summary}{combTMB}(object, ...)

\method{print}{summary.combTMB}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{combTMB}, a result of call
\code{\link{combTMB}(...)}}

\item{\dots}{Currently not used}

\item{x}{A object of class \code{combTMB}}

\item{digits}{minimal number of _significant_ digits, see
\code{\link[base]{print.default}}}
}
\value{
an object of class \code{"summary.combTMB"}, a list with
  components.
}
\description{
Summary of the combTMB models (individual t-tests)
}
\author{
Deoclecio Jardim Amorim  <deocleciojardimamorim@gmail.com>
}
