% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{toenail}
\alias{toenail}
\title{A clinical trial in toenail infection named onychomycosis}
\format{
A data frame with 1908 rows and four variables:
\describe{
\item{idnum}{number of patients}
\item{time}{number of measurements per subject (in time)}
\item{treatn}{two oral treatments (0 and 1)}
\item{onyresp}{response variable: severity of infection, coded as 0
(not severe) or 1 (severe)}
}
}
\usage{
data(toenail, package = "combTMB")
}
\description{
This dataset comes from a randomized, multicenter, parallel-group
study to compare two oral treatments (coded as 0 and 1) for toenail dermatophytic
onychomycosis. For more details see Molenberghs et al. (2010).
}
\references{
Molenberghs, G., Verbeke, G., Demetrio, C. G., and Vieira, A. M. (2010).
A family of generalized linear models for repeated measures with normal
and conjugate random effects. Statistical science, 25(3), 325-347.
}
\keyword{datasets}
